Move showCmpForExistingUser() and hasSupportedCmp() to AppLovinMAX
