+++ src/AppLovinMAX.ts
@@ -6,6 +6,31 @@ const NativeAppLovinMAX = NativeModules.AppLovinMAX;
 
 const VERSION = '6.2.1';
 
+/**
+ * Represents errors for CMP flow.
+ */
+export enum CmpError {
+    /**
+     * Indicates that an unspecified error has occurred.
+     */
+    UNSPECIFIED = -1,
+
+    /**
+     * Indicates that the CMP has not been integrated correctly.
+     */
+    INTEGRATION_ERROR = 1,
+
+    /**
+     * Indicates that the CMP form is unavailable.
+     */
+    FORM_UNAVAILABLE = 2,
+
+    /**
+     * Indicates that the CMP form is not required.
+     */
+    FORM_NOT_REQUIRED = 3,
+}
+
 const initialize = async (sdkKey: string): Promise<Configuration> => {
     return NativeAppLovinMAX.initialize(VERSION, sdkKey);
 };
